requirements.txt
flask
celery
redis
gunicorn
python-docx
pdfplumber
google-generativeai
pillow
python-dotenv